Fix removed np.float alias in gaussian_wave_packet's HDF5 export
np.float was deprecated in NumPy 1.20 and removed in 1.24, so this
would raise AttributeError on any current NumPy install. Changed to
np.float32, matching the dtype already used by the sibling datasets in
the same export block (ProbingPoint, k_vector, timestep).

+++ src/kite/__init__.py
@@ -2138,7 +2138,7 @@ def config_system(lattice, config, calculation, modification=None, **kwargs):
         grpc_p.create_dataset('NumDisorder', data=num_disorder, dtype=np.int32)
         grpc_p.create_dataset('mean_value', data=mean_value, dtype=np.int32)
         grpc_p.create_dataset('ProbingPoint', data=np.asmatrix(np.asarray(probing_points)).astype(np.float32))
-        grpc_p.create_dataset('width', data=width, dtype=np.float)
+        grpc_p.create_dataset('width', data=width, dtype=np.float32)
         grpc_p.create_dataset('spinor', data=np.asmatrix(np.asarray(spinor)).astype(config.type))
         grpc_p.create_dataset('k_vector', data=np.asmatrix(np.asarray(k_vector)), dtype=np.float32)
         grpc_p.create_dataset('timestep', data=timestep, dtype=np.float32)
